backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-genai
PyMuPDF
python-dotenv
